cephadm: move command call functions to call_wrappers.py
Pair-programmed-with: Adam King <adking@redhat.com>
